move not owned summary to its own file
